+++ spotinst/elastigroup_azure_launch_configuration/consts.go
@@ -7,9 +7,9 @@ const (
 )
 
 const (
-	UserData                 commons.FieldName = "user_data"
-	ShutdownScript           commons.FieldName = "shutdown_script"
-	CustomData               commons.FieldName = "custom_data"
+	UserData       commons.FieldName = "user_data"
+	ShutdownScript commons.FieldName = "shutdown_script"
+	//CustomData               commons.FieldName = "custom_data"
 	ManagedServiceIdentities commons.FieldName = "managed_service_identities"
 
 	ResourceGroupName commons.FieldName = "resource_group_name"