+++ CHANGELOG.md
@@ -1,9 +1,10 @@
 ## 1.11.0 (Unreleased)
 
+FEATURES:
+* *New Resource*: `spotinst_ocean_aws_launch_spec`
+
 ENHANCEMENTS:
-* new resource: ocean_aws_launch_spec
 * resource/elastigroup_azure: added `managed_service_identities`
-* resource/elastigroup_azure: added `custom_data`
 * resource/elastigroup_gcp: added `shutdown_script`
 * resource/elastigroup_gcp: added healthcheck fields
 * resource/mrscaler: added `instance_weights`